+++ CODEOWNERS
@@ -16,14 +16,4 @@ requirements*.txt @plugwise/devs
 /scripts/ @bouwew @compatech
 
 # NET specific
-/plugwise/smile.py @bouwew @compatech
-
-# USB specific
-/plugwise/connections/ @brefra
-/plugwise/controller.py @bouwew @brefra
-/plugwise/messages/ @brefra
-/plugwise/nodes/ @brefra
-/plugwise/parser.py @brefra
-/plugwise/stick.py @brefra
-/plugwise/util.py @brefra
-
+/plugwise @bouwew @compatech
