+++ plugwise/exceptions.py
