--- plugwise/stick/nodes/stealth.py
@@ -1,6 +1,0 @@
-"""Plugwise Stealth node object."""
-from ..nodes.circle import PlugwiseCircle
-
-
-class PlugwiseStealth(PlugwiseCircle):
-    """provides interface to the Plugwise Stealth nodes"""
