+++ plugwise/util.py
