+++ pyproject.toml
@@ -4,9 +4,9 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name            = "plugwise"
-version         = "0.30.0a1"
+version         = "0.31.0a1"
 license         = {file = "LICENSE"}
-description     = "Plugwise Smile (Adam/Anna/P1), Stretch and USB (Stick) module for Python 3."
+description     = "Plugwise Smile (Adam/Anna/P1) and Stretch module for Python 3."
 readme          = "README.md"
 keywords        = ["home", "automation", "plugwise", "module"]
 classifiers     = [
@@ -24,7 +24,6 @@ authors = [
 ]
 maintainers = [
         { name = "bouwew"},
-        { name = "brefra"},
         { name = "CoMPaTech" }
 ]
 requires-python = ">=3.9.0"
@@ -206,16 +205,13 @@ disallow_untyped_defs = true
 no_implicit_optional = true
 warn_return_any = true
 warn_unreachable = true
-exclude = [
-    "plugwise/stick/*",
-]
+exclude = []
 
 [tool.coverage.run]
 source = [ "plugwise" ]
 omit= [
     "*/venv/*",
     "setup.py",
-    "plugwise/stick/*",
 ]
 
 [tool.ruff]
@@ -257,9 +253,7 @@ ignore = [
 #    # Ignored due to performance: https://github.com/charliermarsh/ruff/issues/2923
 #    "UP038", # Use `X | Y` in `isinstance` call instead of `(X, Y)`
 
-exclude = [
-    "plugwise/stick/*",
-]
+exclude = []
 
 [tool.ruff.flake8-import-conventions.extend-aliases]
 voluptuous = "vol"
