+++ scripts/complexity.sh
@@ -13,7 +13,7 @@ if [ -f "${my_venv}/bin/activate" ]; then
     echo "-----------------------------"
     echo "Running cyclomatic complexity"
     echo "-----------------------------"
-    PYTHONPATH=$(pwd) radon cc plugwise/smile.py plugwise/helper.py tests/test_smile.py -s -nc --no-assert
+    PYTHONPATH=$(pwd) radon cc plugwise/ tests/ -s -nc --no-assert
 else
     echo "Virtualenv available, bailing out"
     exit 2
