--- tests/test_stick.py
@@ -1,11 +1,0 @@
-# pylint: disable=protected-access
-"""Test Plugwise Stick features."""
-
-import importlib
-
-pw_constants = importlib.import_module("plugwise.stick.constants")
-pw_exceptions = importlib.import_module("plugwise.stick.exceptions")
-pw_stick = importlib.import_module("plugwise.stick")
-
-# No tests available
-assert True
